requirements.txt
Django>=5.0.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
Pillow>=10.2.0
django-taggit>=5.0.1
django-cleanup>=8.0.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
alembic>=1.13.1
python-slugify>=8.0.1
pydantic>=2.6.1
pydantic-settings>=2.1.0
python-dotenv>=1.0.0 